#13226 [fix] 년도별 지표관리, 년도별 현황 페이징처리 제거

+++ src/main/kotlin/co/brainz/itsm/sla/metricYear/dto/MetricYearExcelDto.kt
@@ -11,10 +11,10 @@ import java.io.Serializable
 data class MetricYearExcelDto(
     val metricGroupName: String?,
     val metricName: String?,
-    val minValue: Double?,
-    val maxValue: Double?,
-    val weightValue: Double?,
-    val score: Long?,
+    val minValue: Float?,
+    val maxValue: Float?,
+    val weightValue: Float?,
+    val score: Float?,
     val owner: String?,
     val comment: String?
 ) : Serializable
